refactor(endpoints-overlay): Convert stories to TypeScript.
