+++ packages/endpoints-overlay/package.json
@@ -22,7 +22,6 @@
     "@opentripplanner/location-icon": "^1.4.0",
     "@opentripplanner/core-utils": "^4.5.0",
     "flat": "^5.0.2",
-    "prop-types": "^15.7.2",
     "@styled-icons/fa-solid": "^10.34.0"
   },
   "devDependencies": {
